Changed ReplayBuffer to play better with GPU eager mode, restructured to four files, added Tensor extension, formatted via swift-format.
