+++ Gym/DQN/ReplayBuffer.swift
@@ -1,0 +1,81 @@
+// Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import TensorFlow
+
+struct ReplayBuffer {
+  let capacity: Int
+  let combined: Bool
+
+  @noDerivative var states: [Tensor<Float>] = []
+  @noDerivative var actions: [Tensor<Int32>] = []
+  @noDerivative var rewards: [Tensor<Float>] = []
+  @noDerivative var nextStates: [Tensor<Float>] = []
+  @noDerivative var isDones: [Tensor<Bool>] = []
+  var count: Int { return states.count }
+
+  init(capacity: Int, combined: Bool) {
+    self.capacity = capacity
+    self.combined = combined
+  }
+
+  mutating func append(
+    state: Tensor<Float>,
+    action: Tensor<Int32>,
+    reward: Tensor<Float>,
+    nextState: Tensor<Float>,
+    isDone: Tensor<Bool>
+  ) {
+    if count >= capacity {
+      // Erase oldest SARS if the replay buffer is full
+      states.removeFirst()
+      actions.removeFirst()
+      rewards.removeFirst()
+      nextStates.removeFirst()
+      isDones.removeFirst()
+    }
+    states.append(state)
+    actions.append(action)
+    rewards.append(reward)
+    nextStates.append(nextState)
+    isDones.append(isDone)
+  }
+
+  func sample(batchSize: Int) -> (
+    stateBatch: Tensor<Float>,
+    actionBatch: Tensor<Int32>,
+    rewardBatch: Tensor<Float>,
+    nextStateBatch: Tensor<Float>,
+    isDoneBatch: Tensor<Bool>
+  ) {
+    let indices: Tensor<Int32>
+    if self.combined == true {
+      // Combined Experience Replay
+      let sampledIndices = (0..<batchSize - 1).map { _ in Int32.random(in: 0..<Int32(count)) }
+      indices = Tensor<Int32>(shape: [batchSize], scalars: sampledIndices + [Int32(count) - 1])
+    } else {
+      // Vanilla Experience Replay
+      let sampledIndices = (0..<batchSize).map { _ in Int32.random(in: 0..<Int32(count)) }
+      indices = Tensor<Int32>(shape: [batchSize], scalars: sampledIndices)
+    }
+
+    let stateBatch = Tensor(stacking: states).gathering(atIndices: indices, alongAxis: 0)
+    let actionBatch = Tensor(stacking: actions).gathering(atIndices: indices, alongAxis: 0)
+    let rewardBatch = Tensor(stacking: rewards).gathering(atIndices: indices, alongAxis: 0)
+    let nextStateBatch = Tensor(stacking: nextStates).gathering(atIndices: indices, alongAxis: 0)
+    let isDoneBatch = Tensor(stacking: isDones).gathering(atIndices: indices, alongAxis: 0)
+
+    return (stateBatch, actionBatch, rewardBatch, nextStateBatch, isDoneBatch)
+  }
+}
